feat: Timer model
This class can be use to measure the exeution time of a function,
and stored it in a dictionnary.
To use it just call the start() method before the function you want
to measure by giving an id and then call the stop() method after the
function you want to measure by giving the same id.