feat: use parralle computing for preprocess+index
I use Pool from multiprocessus module.
But since it use a Queue of pickled objects I had to refactor the code
since pickle doesn't accepte lambda or local functions.
For the lemmer and stemmer I just handle this by creating an identity
function, to use instead of lemmer or stemmer if we don't used it.
For the tokenizer I simply remove the possibility to use regular expression
because re.find take 2 param and WordNetTokenizer take only one.
So I prefer loosing this against the add complexity to keep it.

Result: half time to preprocess and index.

TODO: next in order to gain more time I will try to compile the python
code into assembly, maybe using docker to be more portable.
(and we can deploy it in a cloud computing environment)