refactor: handling of xml elements
For know I consider each of them as a separate document in order to
be able to compute statistics. But an error still reamin "math domain"
when computing the score of a query. Also for over query the score is
always null.
We need to check which changer has been made which broke this
computation. Maybe a missing value or a wrong type.